mmo/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.47.0
django-appconf==1.0.5
django-ckeditor==6.2.0
django-js-asset==1.2.2
idna==3.3
oauthlib==3.1.1
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
